+++ src/Sentry.Unity.Editor/Android/GradleSetup.cs
@@ -1,9 +1,6 @@
-using System;
-using System.Collections.Generic;
 using System.IO;
-using System.Linq;
+using System.Text.RegularExpressions;
 using Sentry.Extensibility;
-using UnityEditor.Build;
 
 namespace Sentry.Unity.Editor.Android
 {
@@ -11,11 +8,13 @@ namespace Sentry.Unity.Editor.Android
     {
         private readonly IDiagnosticLogger _logger;
 
-        public const string DependencyScopeName = "dependencies";
-        public const string SdkDependencies = @"implementation(name: 'sentry-android-ndk-release', ext:'aar')
-    implementation(name: 'sentry-android-core-release', ext:'aar')";
-        public static readonly List<string> ScopesToSkip = new() { "buildscript", "pluginManagement" };
+        private const string AndroidMarker = "android {";
+        private const string SdkDependenciesFull = SdkDependencies + "\n}\n\n" + AndroidMarker;
 
+        public const string SdkDependencies = @"dependencies {
+    implementation(name: 'sentry-android-ndk-release', ext:'aar')
+    implementation(name: 'sentry-android-core-release', ext:'aar')";
+        public const string DependenciesAddedMessage = "The Sentry Gradle dependencies have already been added.";
         private readonly string _unityLibraryGradle;
 
         public GradleSetup(IDiagnosticLogger logger, string gradleProjectPath)
@@ -27,111 +26,43 @@ namespace Sentry.Unity.Editor.Android
         public void UpdateGradleProject()
         {
             _logger.LogInfo("Adding Sentry to the gradle project.");
-            var unityLibraryGradleContent = LoadGradleScript(_unityLibraryGradle);
-            unityLibraryGradleContent = InsertIntoScope(unityLibraryGradleContent, DependencyScopeName, SdkDependencies);
-            File.WriteAllText(_unityLibraryGradle, unityLibraryGradleContent);
+            var fileContent = LoadGradleScript(_unityLibraryGradle);
+            fileContent = AddSentryToGradle(fileContent);
+            File.WriteAllText(_unityLibraryGradle, fileContent);
         }
 
         public void ClearGradleProject()
         {
             _logger.LogInfo("Removing Sentry from the gradle project.");
-            var unityLibraryGradleContent = LoadGradleScript(_unityLibraryGradle);
-            unityLibraryGradleContent = RemoveFromGradleContent(unityLibraryGradleContent, SdkDependencies);
-            File.WriteAllText(_unityLibraryGradle, unityLibraryGradleContent);
+            var fileContent = LoadGradleScript(_unityLibraryGradle);
+
+            if (!fileContent.Contains(SdkDependenciesFull))
+            {
+                _logger.LogDebug("The Sentry Gradle dependencies have already been removed.");
+                return;
+            }
+
+            fileContent = fileContent.Replace(SdkDependenciesFull, AndroidMarker);
+            File.WriteAllText(_unityLibraryGradle, fileContent);
         }
 
-        internal string InsertIntoScope(string gradleContent, string scope, string insertion)
+        public string AddSentryToGradle(string fileContent)
         {
-            if (gradleContent.Contains(insertion))
+            if (fileContent.Contains(SdkDependenciesFull))
             {
-                _logger.LogDebug("The gradle file has already been updated. Skipping.");
-                return gradleContent;
+                _logger.LogDebug(DependenciesAddedMessage);
+                return fileContent;
             }
 
-            var lines = gradleContent.Split('\n');
-            var scopeStart = FindBeginningOfScope(lines, scope);
-            if (scopeStart == -1)
-            {
-                throw new BuildFailedException($"Failed to find scope '{scope}'.");
-            }
-
-            var modifiedLines = new List<string>(lines);
-
-            var lineToInsert = string.Empty;
-            var whiteSpaceCount = lines[scopeStart].IndexOf(scope, StringComparison.Ordinal);
-            for (var i = 0; i < whiteSpaceCount; i++)
-            {
-                lineToInsert += " ";
-            }
-
-            lineToInsert += "    " + insertion;
-            var lineOffset = lines[scopeStart].Contains("{") ? 1 : 2; // to make sure we're inside the scope
-            modifiedLines.Insert(scopeStart + lineOffset, lineToInsert);
-
-            return string.Join("\n", modifiedLines.ToArray());
+            var regex = new Regex(Regex.Escape(AndroidMarker));
+            return regex.Replace(fileContent, SdkDependenciesFull, 1);
         }
-
-        private int FindBeginningOfScope(string[] lines, string scope)
-        {
-            for (var i = 0; i < lines.Length; i++)
-            {
-                var line = lines[i];
-                // There are potentially multiple, nested scopes. We cannot add ourselves to the ones listed in `ScopesToSkip`
-                if (ScopesToSkip.Any(line.Contains))
-                {
-                    var startIndex = i;
-
-                    // In case the '{' is on the next line
-                    if (!line.Contains("{") && lines[startIndex + 1].Contains("{"))
-                    {
-                        startIndex += 1;
-                    }
-
-                    i = FindClosingBracket(lines, startIndex);
-                }
-                else if (lines[i].Contains(scope))
-                {
-                    return i;
-                }
-            }
-
-            return -1;
-        }
-
-        private static int FindClosingBracket(string[] lines, int startIndex)
-        {
-            var openBrackets = 0;
-
-            for (var i = startIndex + 1; i < lines.Length; i++)
-            {
-                var line = lines[i];
-                if (line.Contains("{"))
-                {
-                    openBrackets++;
-                }
-
-                if (line.Contains("}"))
-                {
-                    if (openBrackets == 0)
-                    {
-                        return i;
-                    }
-
-                    openBrackets--;
-                }
-            }
-
-            throw new BuildFailedException("Failed to find the closing bracket.");
-        }
-
-        private static string RemoveFromGradleContent(string gradleContent, string toRemove)
-            => gradleContent.Contains(toRemove) ? gradleContent.Replace(toRemove, "") : gradleContent;
 
         internal static string LoadGradleScript(string path)
         {
             if (!File.Exists(path))
             {
-                throw new FileNotFoundException($"Failed to find the gradle config.", path);
+                throw new FileNotFoundException("Failed to find the gradle config.", path);
             }
             return File.ReadAllText(path);
         }